refactored auth folder
moved all files in auth folder into scripts folder
