+++ assets/scripts/events.js
@@ -1,6 +1,6 @@
 'use strict';
 
-const getFormFields = require('../../../lib/get-form-fields');
+const getFormFields = require('../../lib/get-form-fields');
 
 const api = require('./api');
 const ui = require('./ui');
