+++ assets/scripts/ui.js
@@ -1,6 +1,6 @@
 'use strict';
 
-const app = require('../app');
+const app = require('./app');
 
 const signInSuccess = (data) => {
   app.user = data.user;
